vulkan: support mixed-sample render targets
Allow native images to be imported as attachment-only textures while
preserving their declared sampler target, layer range, ownership, and
usage. Use a 64-bit native handle so Vulkan images are not truncated on
32-bit clients, and defer attachment-only layout transitions until the
first render-pass use.

Extend Filament's subpass interface with multisampled input attachments
and complete Vulkan descriptor metadata, copy-on-write rebinding, image
layouts, and input-attachment synchronization.

Represent mixed-sample render passes explicitly. Each color attachment
has its own sample count; a multisampled attachment can become
input-only in subpass one, which renders at one sample with no depth.
Native depth resolve is attached to the last subpass that uses depth.
Cache keys include all new state.

Add an explicit shared-context depth-resolve capability. Native-resolve
depth sidecars avoid sampled usage, use transient memory when available,
are discarded after resolve, and are cached per texture view, sample
count, and resolve mode. Resolve only aspects present in the format.

Do not infer transient attachment usage for imported images. Their
creation flags belong to the external owner; marking attachment-only XR
images transient selects DONT_CARE stores and loses on-tile subpass
output.

Expose a backend query for offscreen RenderTarget sample counts instead
of relying on default-swapchain MSAA support. Add compiler coverage for
subpassInputMS and keep material debug serialization in sync.

+++ filament/backend/src/opengl/OpenGLDriver.h
@@ -398,9 +398,9 @@ private:
     void createTextureViewSwizzleCommon(Handle<HwTexture> th, Handle<HwTexture> srch,
             TextureSwizzle r, TextureSwizzle g, TextureSwizzle b,
             TextureSwizzle a, utils::ImmutableCString&& tag);
-    void importTextureCommon(OpenGLState& gl, Handle<HwTexture> th, intptr_t id, SamplerType target, uint8_t levels,
-            TextureFormat format, uint8_t samples, uint32_t width, uint32_t height, uint32_t depth,
-            TextureUsage usage, utils::ImmutableCString&& tag);
+    void importTextureCommon(OpenGLState& gl, Handle<HwTexture> th, uint64_t id, SamplerType target,
+            uint8_t levels, TextureFormat format, uint8_t samples, uint32_t width, uint32_t height,
+            uint32_t depth, TextureUsage usage, utils::ImmutableCString&& tag);
     void createBufferObjectCommon(OpenGLState& gl, Handle<HwBufferObject> boh, uint32_t byteCount,
             BufferObjectBinding bindingType, BufferUsage usage, utils::ImmutableCString&& tag);
     void setVertexBufferObjectCommon(Handle<HwVertexBuffer> vbh, uint32_t index,
